BigQueryBaseConfig refactor
