+++ docs/BigQueryExecute-action.md
@@ -49,7 +49,8 @@ cache that will be flushed whenever tables in the query are modified.
 **Job Location**: Location of the job. It must match the location of the dataset specified in the query.
 
 **Encryption Key Name**: Used to encrypt data written to any dataset or table created by the plugin.
-If the dataset or table already exists, this is ignored.
+If the dataset or table already exists, this is ignored. More information can be found at 
+https://cloud.google.com/data-fusion/docs/how-to/customer-managed-encryption-keys"
 
 **Row As Arguments**: Row as arguments. For example, if the query is 'select min(id) as min_id, max(id) as max_id from my_dataset.my_table',
 an arguments for 'min_id' and 'max_id' will be set based on the query results. Plugins further down the pipeline can then
