+++ docs/BigQueryPushdownEngine-sqlengine.md
@@ -49,7 +49,8 @@ bucket will be created and then deleted after the run finishes.
 if the dataset or temporary bucket already exist.
 
 **Encryption Key Name**: Used to encrypt data written to any bucket, dataset, or table created by the plugin.
-If the bucket, dataset, or table already exists, this is ignored.
+If the bucket, dataset, or table already exists, this is ignored. More information can be found at 
+https://cloud.google.com/data-fusion/docs/how-to/customer-managed-encryption-keys"
 
 **Retain BigQuery tables after completion**: By Default, the temporary BigQuery tables used to execute operations
 will be deleted after execution is completed. Use this setting to override the default behavior and retain tables.
