+++ docs/GCS-batchsink.md
@@ -53,7 +53,8 @@ The delimiter will be ignored if the format is anything other than 'delimited'.
 **Location:** The location where the gcs bucket will get created. This value is ignored if the bucket already exists.
 
 **Encryption Key Name:** Used to encrypt data written to any bucket created by the plugin.
-If the bucket already exists, this is ignored.
+If the bucket already exists, this is ignored. More information can be found at 
+https://cloud.google.com/data-fusion/docs/how-to/customer-managed-encryption-keys"
 
 **Content Type:** The Content Type entity is used to indicate the media type of the resource.
 Defaults to 'application/octet-stream'. The following table shows valid content types for each format.
