+++ docs/GCSBucketCreate-action.md
@@ -37,7 +37,8 @@ It can be found on the Dashboard in the Google Cloud Platform Console.
 **Location:** The location where the gcs buckets will get created. This value is ignored if the bucket already exists.
 
 **Encryption Key Name:** Used to encrypt data written to any bucket created by the plugin.
-If the bucket already exists, this is ignored.
+If the bucket already exists, this is ignored. More information can be found at 
+https://cloud.google.com/data-fusion/docs/how-to/customer-managed-encryption-keys"
 
 **Service Account**  - service account key used for authorization
