+++ docs/GCSMultiFiles-batchsink.md
@@ -63,7 +63,8 @@ The delimiter will be ignored if the format is anything other than 'delimited'.
 **Location:** The location where the gcs buckets will get created. This value is ignored if the bucket already exists.
 
 **Encryption Key Name:** Used to encrypt data written to any bucket created by the plugin.
-If the bucket already exists, this is ignored.
+If the bucket already exists, this is ignored. More information can be found at 
+https://cloud.google.com/data-fusion/docs/how-to/customer-managed-encryption-keys"
 
 **Service Account**  - service account key used for authorization
