+++ src/main/java/io/cdap/plugin/gcp/bigquery/sink/AbstractBigQuerySinkConfig.java
@@ -15,24 +15,13 @@
  */
 package io.cdap.plugin.gcp.bigquery.sink;
 
-import com.google.auth.Credentials;
-import com.google.cloud.ServiceOptions;
-import com.google.cloud.bigquery.BigQuery;
-import com.google.cloud.bigquery.Dataset;
-import com.google.cloud.bigquery.DatasetId;
 import com.google.cloud.bigquery.JobInfo;
-import com.google.cloud.bigquery.TableId;
 import com.google.cloud.kms.v1.CryptoKeyName;
-import com.google.cloud.storage.Bucket;
-import com.google.cloud.storage.Storage;
-import com.google.cloud.storage.StorageException;
-import com.google.common.base.Strings;
 import com.google.common.collect.ImmutableSet;
 import io.cdap.cdap.api.annotation.Description;
 import io.cdap.cdap.api.annotation.Macro;
 import io.cdap.cdap.api.annotation.Name;
 import io.cdap.cdap.api.data.schema.Schema;
-import io.cdap.cdap.api.plugin.PluginConfig;
 import io.cdap.cdap.etl.api.FailureCollector;
 import io.cdap.plugin.common.ConfigUtil;
 import io.cdap.plugin.common.Constants;
@@ -41,8 +30,6 @@ import io.cdap.plugin.gcp.bigquery.common.BigQueryBaseConfig;
 import io.cdap.plugin.gcp.bigquery.connector.BigQueryConnectorConfig;
 import io.cdap.plugin.gcp.bigquery.util.BigQueryUtil;
 import io.cdap.plugin.gcp.common.CmekUtils;
-import io.cdap.plugin.gcp.common.GCPUtils;
-import io.cdap.plugin.gcp.gcs.GCSPath;
 
 import java.util.Collections;
 import java.util.Map;
@@ -52,7 +39,7 @@ import javax.annotation.Nullable;
 /**
  * Base class for Big Query batch sink configs.
  */
-public abstract class AbstractBigQuerySinkConfig extends PluginConfig {
+public abstract class AbstractBigQuerySinkConfig extends BigQueryBaseConfig {
   public static final Set<Schema.Type> SUPPORTED_TYPES =
     ImmutableSet.of(Schema.Type.INT, Schema.Type.LONG, Schema.Type.STRING, Schema.Type.FLOAT, Schema.Type.DOUBLE,
                     Schema.Type.BOOLEAN, Schema.Type.BYTES, Schema.Type.ARRAY, Schema.Type.RECORD);
@@ -62,28 +49,10 @@ public abstract class AbstractBigQuerySinkConfig extends PluginConfig {
   private static final String NAME_GCS_CHUNK_SIZE = "gcsChunkSize";
   protected static final String NAME_UPDATE_SCHEMA = "allowSchemaRelaxation";
   private static final String SCHEME = "gs://";
-  protected static final String NAME_DATASET = "dataset";
-  private static final String NAME_BUCKET = "bucket";
-  private static final String NAME_CMEK_KEY = "cmekKey";
 
   @Name(Constants.Reference.REFERENCE_NAME)
   @Description(Constants.Reference.REFERENCE_NAME_DESCRIPTION)
   protected String referenceName;
-
-  @Name(NAME_DATASET)
-  @Macro
-  @Description("The dataset to write to. A dataset is contained within a specific project. "
-    + "Datasets are top-level containers that are used to organize and control access to tables and views.")
-  protected String dataset;
-
-  @Name(NAME_BUCKET)
-  @Macro
-  @Nullable
-  @Description("The Google Cloud Storage bucket to store temporary data in. "
-    + "Cloud Storage data will be deleted after it is loaded into BigQuery. "
-    + "If it is not provided, a unique bucket will be automatically created and then deleted after the run finishes. "
-    + "The service account must have permission to create buckets in the configured project.")
-  protected String bucket;
 
   @Name(NAME_GCS_CHUNK_SIZE)
   @Macro
@@ -112,24 +81,9 @@ public abstract class AbstractBigQuerySinkConfig extends PluginConfig {
     "This value is ignored if the dataset or temporary bucket already exist.")
   protected String location;
 
-  @Name(NAME_CMEK_KEY)
-  @Macro
-  @Nullable
-  @Description("The GCP customer managed encryption key (CMEK) name used to encrypt data written to " +
-    "any bucket, dataset, or table created by the plugin. If the bucket, dataset, or table already exists, " +
-    "this is ignored.")
-  protected String cmekKey;
-
-  @Name(ConfigUtil.NAME_USE_CONNECTION)
-  @Nullable
-  @Description("Whether to use an existing connection.")
-  private Boolean useConnection;
-
-  @Name(ConfigUtil.NAME_CONNECTION)
-  @Macro
-  @Nullable
-  @Description("The existing connection to use.")
-  protected BigQueryConnectorConfig connection;
+  public AbstractBigQuerySinkConfig(BigQueryConnectorConfig connection, String dataset, String cmekKey, String bucket) {
+    super(connection, dataset, cmekKey, bucket);
+  }
 
   public String getReferenceName() {
     return referenceName;
@@ -166,7 +120,7 @@ public abstract class AbstractBigQuerySinkConfig extends PluginConfig {
   public void validate(FailureCollector collector) {
     validate(collector, Collections.emptyMap());
   }
-  
+
   public void validate(FailureCollector collector, Map<String, String> arguments) {
     IdUtils.validateReferenceName(referenceName, collector);
     ConfigUtil.validateConnection(this, useConnection, connection, collector);
@@ -194,97 +148,8 @@ public abstract class AbstractBigQuerySinkConfig extends PluginConfig {
     validateCmekKeyLocation(cmekKeyName, null, location, failureCollector);
   }
 
-  public String getDataset() {
-    return dataset;
-  }
-
   public String getDatasetProject() {
     return connection == null ? null : connection.getDatasetProject();
   }
 
-  public String getProject() {
-    if (connection == null) {
-      throw new IllegalArgumentException(
-        "Could not get project information, connection should not be null!");
-    }
-    return connection.getProject();
-  }
-
-  @Nullable
-  public String tryGetProject() {
-    return connection == null ? null : connection.tryGetProject();
-  }
-
-  @Nullable
-  public String getServiceAccount() {
-    return connection == null ? null : connection.getServiceAccount();
-  }
-
-  @Nullable
-  public Boolean isServiceAccountFilePath() {
-    return connection == null ? null : connection.isServiceAccountFilePath();
-  }
-
-  @Nullable
-  public String getServiceAccountType() {
-    return connection == null ? null : connection.getServiceAccountType();
-  }
-
-  @Nullable
-  public String getBucket() {
-    if (bucket != null) {
-      bucket = bucket.trim();
-      if (bucket.isEmpty()) {
-        return null;
-      }
-      // remove the gs:// scheme from the bucket name
-      if (bucket.startsWith(SCHEME)) {
-        bucket = bucket.substring(SCHEME.length());
-      }
-    }
-    return bucket;
-  }
-
-  /* returns the bucket if it exists otherwise null.
-   */
-  private Bucket getBucketIfExists(Storage storage, String bucketName) {
-    Bucket bucket = null;
-    try {
-      bucket = storage.get(bucketName);
-    } catch (StorageException e) {
-      // If there is an exception getting bucket information during config validation, it will be ignored because
-      // service account used can be different.
-    }
-    return bucket;
-  }
-
-  public void validateCmekKeyLocation(@Nullable CryptoKeyName cmekKeyName, @Nullable String tableName,
-                                      @Nullable String location, FailureCollector failureCollector) {
-    if (cmekKeyName == null || containsMacro(NAME_DATASET) || connection == null || !connection.canConnect()
-      || containsMacro(NAME_BUCKET)) {
-      return;
-    }
-    String datasetProjectId = connection.getDatasetProject();
-    String datasetName = getDataset();
-    DatasetId datasetId = DatasetId.of(datasetProjectId, datasetName);
-    TableId tableId = tableName == null ? null : TableId.of(datasetProjectId, datasetName, tableName);
-    Credentials credentials = connection.getCredentials(failureCollector);
-    BigQuery bigQuery = GCPUtils.getBigQuery(connection.getProject(), credentials);
-    Storage storage = GCPUtils.getStorage(connection.getProject(), credentials);
-    if (bigQuery == null || storage == null) {
-      return;
-    }
-    String bucketName = getBucket();
-    Bucket bucket = bucketName == null ? null : getBucketIfExists(storage, bucketName);
-    // if bucket exists then dataset and table will be created in location of bucket if they do not exist.
-    location = bucket == null ? location : bucket.getLocation();
-    Dataset dataset = CmekUtils.validateCmekKeyAndDatasetOrTableLocation(bigQuery, datasetId, tableId, cmekKeyName,
-                                                                         location, failureCollector);
-    if (bucket == null && dataset != null) {
-      // if dataset exists then bucket will be created in the location of dataset.
-      location = dataset.getLocation();
-      GCSPath gcsPath = Strings.isNullOrEmpty(bucketName) ? null : GCSPath.from(bucketName);
-      CmekUtils.validateCmekKeyAndBucketLocation(storage, gcsPath, cmekKeyName, location, failureCollector);
-    }
-  }
 }
