+++ src/main/java/io/cdap/plugin/gcp/bigquery/sink/BigQueryMultiSinkConfig.java
@@ -19,6 +19,7 @@ import com.google.common.base.Strings;
 import io.cdap.cdap.api.annotation.Description;
 import io.cdap.cdap.api.annotation.Macro;
 import io.cdap.cdap.api.annotation.Name;
+import io.cdap.plugin.gcp.bigquery.connector.BigQueryConnectorConfig;
 
 import javax.annotation.Nullable;
 
@@ -42,11 +43,53 @@ public class BigQueryMultiSinkConfig extends AbstractBigQuerySinkConfig {
     "arguments will be processed. If enabled, all records will be written as-is.")
   private Boolean allowFlexibleSchema;
 
+  private BigQueryMultiSinkConfig(BigQueryConnectorConfig connection, String dataset, String cmekKey, String bucket) {
+    super(connection, dataset, cmekKey, bucket);
+  }
+
+  public static BigQueryMultiSinkConfig.Builder builder() {
+    return new BigQueryMultiSinkConfig.Builder();
+  }
+
   public String getSplitField() {
     return Strings.isNullOrEmpty(splitField) ? SPLIT_FIELD_DEFAULT : splitField;
   }
 
   public Boolean getAllowFlexibleSchema() {
     return allowFlexibleSchema != null ? allowFlexibleSchema : false;
+  }
+
+  /**
+   * BigQuery MultiSink configuration builder.
+   */
+  public static class Builder {
+    private BigQueryConnectorConfig connection;
+    private String dataset;
+    private String cmekKey;
+    private String bucket;
+
+    public BigQueryMultiSinkConfig.Builder setConnection(@Nullable BigQueryConnectorConfig connection) {
+      this.connection = connection;
+      return this;
+    }
+
+    public BigQueryMultiSinkConfig.Builder setDataset(@Nullable String dataset) {
+      this.dataset = dataset;
+      return this;
+    }
+
+    public BigQueryMultiSinkConfig.Builder setCmekKey(@Nullable String cmekKey) {
+      this.cmekKey = cmekKey;
+      return this;
+    }
+
+    public BigQueryMultiSinkConfig.Builder setBucket(@Nullable String bucket) {
+      this.bucket = bucket;
+      return this;
+    }
+
+    public BigQueryMultiSinkConfig build() {
+      return new BigQueryMultiSinkConfig(connection, dataset, cmekKey, bucket);
+    }
   }
 }
