+++ src/main/java/io/cdap/plugin/gcp/bigquery/source/BigQuerySourceConfig.java
@@ -17,7 +17,6 @@
 package io.cdap.plugin.gcp.bigquery.source;
 
 import com.google.auth.Credentials;
-import com.google.auth.oauth2.ServiceAccountCredentials;
 import com.google.cloud.ServiceOptions;
 import com.google.cloud.bigquery.BigQuery;
 import com.google.cloud.bigquery.Dataset;
@@ -32,12 +31,11 @@ import io.cdap.cdap.api.annotation.Description;
 import io.cdap.cdap.api.annotation.Macro;
 import io.cdap.cdap.api.annotation.Name;
 import io.cdap.cdap.api.data.schema.Schema;
-import io.cdap.cdap.api.plugin.PluginConfig;
-import io.cdap.cdap.etl.api.Arguments;
 import io.cdap.cdap.etl.api.FailureCollector;
 import io.cdap.plugin.common.ConfigUtil;
 import io.cdap.plugin.common.Constants;
 import io.cdap.plugin.common.IdUtils;
+import io.cdap.plugin.gcp.bigquery.common.BigQueryBaseConfig;
 import io.cdap.plugin.gcp.bigquery.connector.BigQueryConnectorConfig;
 import io.cdap.plugin.gcp.bigquery.util.BigQueryUtil;
 import io.cdap.plugin.gcp.common.CmekUtils;
@@ -53,16 +51,14 @@ import javax.annotation.Nullable;
 /**
  * Holds configuration required for configuring {@link BigQuerySource}.
  */
-public final class BigQuerySourceConfig extends PluginConfig {
+public final class BigQuerySourceConfig extends BigQueryBaseConfig {
   private static final String SCHEME = "gs://";
   private static final String WHERE = "WHERE";
   public static final Set<Schema.Type> SUPPORTED_TYPES =
     ImmutableSet.of(Schema.Type.LONG, Schema.Type.STRING, Schema.Type.DOUBLE, Schema.Type.BOOLEAN, Schema.Type.BYTES,
                     Schema.Type.ARRAY, Schema.Type.RECORD);
 
-  public static final String NAME_DATASET = "dataset";
   public static final String NAME_TABLE = "table";
-  public static final String NAME_BUCKET = "bucket";
   public static final String NAME_SCHEMA = "schema";
   public static final String NAME_PARTITION_FROM = "partitionFrom";
   public static final String NAME_PARTITION_TO = "partitionTo";
@@ -70,17 +66,11 @@ public final class BigQuerySourceConfig extends PluginConfig {
   public static final String NAME_ENABLE_QUERYING_VIEWS = "enableQueryingViews";
   public static final String NAME_VIEW_MATERIALIZATION_PROJECT = "viewMaterializationProject";
   public static final String NAME_VIEW_MATERIALIZATION_DATASET = "viewMaterializationDataset";
-  public static final String NAME_CMEK_KEY = "cmekKey";
 
   @Name(Constants.Reference.REFERENCE_NAME)
   @Description("This will be used to uniquely identify this source for lineage, annotating metadata, etc.")
   public String referenceName;
 
-  @Name(NAME_DATASET)
-  @Macro
-  @Description("The dataset the table belongs to. A dataset is contained within a specific project. "
-    + "Datasets are top-level containers that are used to organize and control access to tables and views.")
-  private String dataset;
 
   @Name(NAME_TABLE)
   @Macro
@@ -88,15 +78,6 @@ public final class BigQuerySourceConfig extends PluginConfig {
     + "Each record is composed of columns (also called fields). "
     + "Every table is defined by a schema that describes the column names, data types, and other information.")
   private String table;
-
-  @Name(NAME_BUCKET)
-  @Macro
-  @Nullable
-  @Description("The Google Cloud Storage bucket to store temporary data in. "
-    + "Temporary data will be deleted after it has been read. "
-    + "If it is not provided, a unique bucket will be created and then deleted after the run finishes. "
-    + "The service account must have permission to create buckets in the configured project.")
-  private String bucket;
 
   @Name(NAME_SCHEMA)
   @Macro
@@ -146,46 +127,8 @@ public final class BigQuerySourceConfig extends PluginConfig {
     + "Defaults to the same dataset in which the view is located.")
   private String viewMaterializationDataset;
 
-  @Name(ConfigUtil.NAME_USE_CONNECTION)
-  @Nullable
-  @Description("Whether to use an existing connection.")
-  private Boolean useConnection;
-
-  @Name(ConfigUtil.NAME_CONNECTION)
-  @Macro
-  @Nullable
-  @Description("The existing connection to use.")
-  private BigQueryConnectorConfig connection;
-
-  @Name(NAME_CMEK_KEY)
-  @Macro
-  @Nullable
-  @Description("The GCP customer managed encryption key (CMEK) name used to encrypt data written to " +
-    "any bucket created by the plugin. If the bucket already exists, this is ignored.")
-  protected String cmekKey;
-
-
-  public String getDataset() {
-    return dataset;
-  }
-
   public String getTable() {
     return table;
-  }
-
-  @Nullable
-  public String getBucket() {
-    if (bucket != null) {
-      bucket = bucket.trim();
-      // remove the gs:// scheme from the bucket name
-      if (bucket.startsWith(SCHEME)) {
-        bucket = bucket.substring(SCHEME.length());
-      }
-      if (bucket.isEmpty()) {
-        return null;
-      }
-    }
-    return bucket;
   }
 
   public String getDatasetProject() {
@@ -193,34 +136,6 @@ public final class BigQuerySourceConfig extends PluginConfig {
       return ServiceOptions.getDefaultProjectId();
     }
     return connection.getDatasetProject();
-  }
-
-  public String getProject() {
-    if (connection == null) {
-      throw new IllegalArgumentException(
-        "Could not get project information, connection should not be null!");
-    }
-    return connection.getProject();
-  }
-
-  @Nullable
-  public String tryGetProject() {
-    return connection == null ? null : connection.tryGetProject();
-  }
-
-  @Nullable
-  public String getServiceAccount() {
-    return connection == null ? null : connection.getServiceAccount();
-  }
-
-  @Nullable
-  public Boolean isServiceAccountFilePath() {
-    return connection == null ? null : connection.isServiceAccountFilePath();
-  }
-
-  @Nullable
-  public String getServiceAccountType() {
-    return connection == null ? null : connection.getServiceAccountType();
   }
 
   public void validate(FailureCollector collector) {
@@ -370,37 +285,11 @@ public final class BigQuerySourceConfig extends PluginConfig {
       connection != null && connection.canConnect();
   }
 
-  /**
-   * Return true if the service account is set to auto-detect but it can't be fetched from the environment.
-   * This shouldn't result in a deployment failure, as the credential could be detected at runtime if the pipeline
-   * runs on dataproc. This should primarily be used to check whether certain validation logic should be skipped.
-   *
-   * @return true if the service account is set to auto-detect but it can't be fetched from the environment.
-   */
-  public boolean autoServiceAccountUnavailable() {
-    if (connection == null || connection.getServiceAccountFilePath() == null &&
-      connection.isServiceAccountFilePath()) {
-      try {
-        ServiceAccountCredentials.getApplicationDefault();
-      } catch (IOException e) {
-        return true;
-      }
+    private BigQuerySourceConfig(@Nullable BigQueryConnectorConfig connection, @Nullable String dataset,
+                                @Nullable String cmekKey, @Nullable String bucket, @Nullable String table) {
+      super(connection, dataset, cmekKey, bucket);
+      this.table = table;
     }
-    return false;
-  }
-
-  public BigQueryConnectorConfig getConnection() {
-    return connection;
-  }
-
-  private BigQuerySourceConfig(@Nullable BigQueryConnectorConfig connection, @Nullable String dataset,
-                               @Nullable String cmekKey, @Nullable String bucket, @Nullable String table) {
-    this.connection = connection;
-    this.dataset = dataset;
-    this.cmekKey = cmekKey;
-    this.bucket = bucket;
-    this.table = table;
-  }
 
   public static Builder builder() {
      return new Builder();
